feat: combine asset serving function #167
