+++ dadi/lib/index.js
@@ -39,7 +39,7 @@ var Controller = require(path.join(__dirname, '/controller'))
 var forceDomain = require(path.join(__dirname, '/controller/forceDomain'))
 var help = require(path.join(__dirname, '/help'))
 var Middleware = require(path.join(__dirname, '/middleware'))
-var servePublic = require(path.join(__dirname, '/middleware/servePublic'))
+var servePublic = require(path.join(__dirname, '/view/public'))
 var monitor = require(path.join(__dirname, '/monitor'))
 var Page = require(path.join(__dirname, '/page'))
 var Preload = require(path.resolve(path.join(__dirname, 'datasource/preload')))
@@ -158,7 +158,9 @@ Server.prototype.start = function (done) {
   if (config.get('headers.useGzipCompression')) app.use(compress())
 
   // init main public path for static files
-  if (options.publicPath) servePublic(app, options.publicPath, [])
+  if (options.publicPath) {
+    app.use(servePublic.middleware(options.publicPath, []))
+  }
 
   // init virtual host public paths
   _.each(config.get('virtualHosts'), (virtualHost, key) => {
@@ -171,13 +173,13 @@ Server.prototype.start = function (done) {
       var hostConfig = JSON.parse(fs.readFileSync(hostConfigFile).toString())
       var hostOptions = this.loadPaths(hostConfig.paths)
 
-      servePublic(app, hostOptions.publicPath, virtualHost.hostnames)
+      servePublic.middleware(app, hostOptions.publicPath, virtualHost.hostnames)
     }
   })
 
   // add debug files to static paths
   if (config.get('debug')) {
-    servePublic(
+    servePublic.middleware(
       app,
       options.workspacePath + '/debug' ||
         path.join(__dirname, '/../../workspace/debug'),
